fix(prd): support float priority values in markdown parser
LLMs sometimes generate decimal priorities (e.g. 0.1, 1.5). The parser
used int/%d which truncated these values. Switch Priority field to
float64 and parse with %g to handle both integer and decimal values.

+++ internal/prd/prd_test.go
@@ -333,7 +333,7 @@ func TestPRD_NextStoryContext_PromptSizeUnder10KB(t *testing.T) {
 			Title:              fmt.Sprintf("Story %d with a reasonably long title for realism", i+1),
 			Description:        "This is a description that is moderately long to simulate realistic PRD content for testing purposes.",
 			AcceptanceCriteria: []string{"Criterion A", "Criterion B", "Criterion C"},
-			Priority:           i + 1,
+			Priority:           float64(i + 1),
 			Passes:             i > 0,
 		}
 	}
